fix issues with pr #86
