+++ .gitignore
@@ -16,3 +16,4 @@ logs/
 .vscode/
 talaria
 *.so
+local/
