+++ internal/storage/writer/azure/azure_test.go
@@ -1,16 +1,44 @@
 package azure
 
 import (
+	"bytes"
 	"os"
 	"strings"
 	"testing"
 
+	eorc "github.com/crphang/orc"
 	"github.com/kelindar/talaria/internal/encoding/block"
+	"github.com/kelindar/talaria/internal/encoding/orc"
+	"github.com/kelindar/talaria/internal/encoding/typeof"
 	"github.com/kelindar/talaria/internal/monitor"
 	"github.com/kelindar/talaria/internal/monitor/logging"
 	"github.com/kelindar/talaria/internal/monitor/statsd"
 	"github.com/stretchr/testify/assert"
 )
+
+// Create a base block for testing purpose
+func blockBase() ([]block.Block, error) {
+	schema := typeof.Schema{
+		"col0": typeof.String,
+		"col1": typeof.Int64,
+		"col2": typeof.Float64,
+	}
+	orcSchema, _ := orc.SchemaFor(schema)
+
+	orcBuffer1 := &bytes.Buffer{}
+	writer, _ := eorc.NewWriter(orcBuffer1,
+		eorc.SetSchema(orcSchema))
+	_ = writer.Write("eventName", 1, 1.0)
+	_ = writer.Close()
+
+	apply := block.Transform(nil)
+
+	block, err := block.FromOrcBy(orcBuffer1.Bytes(), "col0", nil, apply)
+	if err != nil {
+		return nil, err
+	}
+	return block, nil
+}
 
 func TestWriter(t *testing.T) {
 	os.Setenv("AZURE_STORAGE_ACCOUNT", "golangrocksonazure")
@@ -20,7 +48,7 @@ func TestWriter(t *testing.T) {
 	assert.NotNil(t, c)
 	assert.NoError(t, err)
 
-	b, err := block.Base()
+	b, err := blockBase()
 	assert.Nil(t, err)
 	assert.NotPanics(t, func() {
 		err := c.Write([]byte("abc"), b)
